requirements.txt
psycopg2-binary==2.8.6
python-dotenv==0.17.1

loguru~=0.5.3
Django~=3.2.6
django-model-utils~=4.0.0
tqdm~=4.62.0
pytz~=2021.1